requirements.txt
requests==2.31.0
urllib3==1.26.16
openai>=1.0.0
tqdm>=4.66.0
nltk>=3.8.1
sentence-transformers>=2.2.2
transformers>=4.40.0
faiss-cpu>=1.7.4
sentencepiece>=0.1.99
protobuf>=3.20.0,<5.0.0
chromadb>=0.4.20
llama-index>=0.10.20